Refactor bisection by removing the padding and adding the padding to the use-chart-hover hook Display tooltip above the chart to keep in line with other charts
